seperate modal from createNecessity
