+++ orange/package.json
@@ -14,6 +14,7 @@
     "@types/node": "^12.0.0",
     "@types/react": "^16.9.0",
     "@types/react-dom": "^16.9.0",
+    "@types/react-modal": "^3.10.6",
     "@types/react-redux": "^7.1.9",
     "@types/react-router": "^5.1.8",
     "@types/react-router-dom": "^5.1.5",
@@ -30,6 +31,7 @@
     "react-dom": "^16.13.1",
     "react-hook-form": "^6.11.2",
     "react-icons": "^3.10.0",
+    "react-modal": "^3.12.1",
     "react-redux": "^7.2.0",
     "react-router-dom": "^5.1.5",
     "react-scripts": "3.4.1",
