+++ orange/src/components/Necessity/NecessityList/NecessityList.tsx
@@ -1,10 +1,11 @@
 import React, { useState } from 'react';
 import './NecessityList.css';
 import { useDispatch } from 'react-redux';
+import Modal from 'react-modal';
 import { Necessity, Place } from '../../../api';
 import NecessityItem from '../NecessityItem/NecessityItem';
 import NecessityCounter from '../NecessityCounter/NecessityCounter';
-import NecessityCreateModal from '../NecessityCreateModal/NecessityCreateModal';
+import NecessityCreate from '../NecessityCreate/NecessityCreate';
 import { createNecessityPlace } from '../../../store/actions/necessity/necessity';
 
 interface Props {
@@ -12,6 +13,7 @@ interface Props {
 }
 
 function NecessityList(props: Props) {
+  const [isModalOpen, setModalOpen] = useState(false);
   const dispatch = useDispatch();
   const { place } = props;
 
@@ -28,7 +30,15 @@ function NecessityList(props: Props) {
 
   return (
     <>
-      <NecessityCreateModal onCreateNecessityPlace={onCreateNecessityPlace} placeId={place.id} />
+      <Modal
+        isOpen={isModalOpen}
+        onRequestClose={() => setModalOpen(false)}
+        contentLabel="Example Modal"
+      >
+        <NecessityCreate onCreateNecessityPlace={onCreateNecessityPlace} placeId={place.id} />
+      </Modal>
+
+      <button type="button" onClick={() => setModalOpen(true)}>만들기</button>
       { place.necessities.length
         ? place.necessities.map((necessity: Necessity) => <NecessityItem necessity={necessity} />)
         : (
