+++ orange/src/components/index.tsx
@@ -1,6 +1,6 @@
 import Header from './Header/Header';
 import Login from './Login/Login';
-import NecessityCreateModal from './Necessity/NecessityCreateModal/NecessityCreateModal';
+import NecessityCreate from './Necessity/NecessityCreate/NecessityCreate';
 import PlaceBox from './Necessity/PlaceBox/PlaceBox';
 import NecessityItem from './Necessity/NecessityItem/NecessityItem';
 import NecessityList from './Necessity/NecessityList/NecessityList';
@@ -11,7 +11,7 @@ import LogList from './LogList/LogList';
 export {
   Header,
   Login,
-  NecessityCreateModal,
+  NecessityCreate,
   PlaceBox,
   NecessityItem,
   NecessityList,
